requirements.txt
amqp==5.2.0asgiref==3.8.1beautifulsoup4==4.12.3billiard==4.2.0Brotli==1.1.0bs4==0.0.2celery==5.4.0certifi==2024.7.4cffi==1.17.0charset-normalizer==3.3.2click==8.1.7click-didyoumean==0.3.1click-plugins==1.1.1click-repl==0.3.0colorama==0.4.6cron-descriptor==1.4.3cryptography==43.0.0Django==5.0.8django-allauth==64.1.0django-celery-beat==2.6.0django-cors-headers==4.4.0django-filter==24.2django-timezone-field==7.0djangorestframework==3.15.2drf-yasg==1.21.7gunicorn==23.0.0idna==3.7inflection==0.5.1kombu==5.4.0mutagen==1.47.0packaging==24.1pillow==10.4.0prompt_toolkit==3.0.47psycopg2-binary==2.9.9pycparser==2.22pycryptodomex==3.20.0PyJWT==2.9.0pypng==0.20220715.0python-crontab==3.2.0python-dateutil==2.9.0.post0pytz==2024.1PyYAML==6.0.2qrcode==7.4.2redis==5.0.8requests==2.32.3setuptools==73.0.0six==1.16.0soupsieve==2.5sqlparse==0.5.1typing_extensions==4.12.2tzdata==2024.1uritemplate==4.1.1urllib3==2.2.2vine==5.1.0wcwidth==0.2.13websockets==12.0yt-dlp==2024.8.1